답글 추가를 할 때도 hook에서 query 최신화를 하게끔 변경
