+++ src/comment/api/comment.ts
@@ -1,7 +1,28 @@
-import { collection, getDocs } from 'firebase/firestore';
+import { collection, getDocs, addDoc, doc, serverTimestamp } from 'firebase/firestore';
 import { firestore } from '@/firebase';
 import { Comment } from '@/comment/model/Comment';
 import { buildNotInQuery } from '@/user/api/user';
+
+/**
+ * 댓글 추가 (순수 데이터 mutation 함수)
+ */
+export async function addCommentToPost(
+  boardId: string,
+  postId: string,
+  content: string,
+  userId: string,
+  userName: string,
+  userProfileImage: string,
+) {
+  const postRef = doc(firestore, `boards/${boardId}/posts/${postId}`);
+  await addDoc(collection(postRef, 'comments'), {
+    content,
+    userId,
+    userName,
+    userProfileImage,
+    createdAt: serverTimestamp(),
+  });
+}
 
 /**
  * 댓글 목록을 한 번만 가져오는 함수 (blockedByUsers 서버사이드 필터링 지원)
